implement a timeout
automatically kills the imap proecess (and therefore the connection) if the
server takes too long.

controlled with a option entry of command_timeout, which should be a
positive integer to create a timeout. anything else and the imap
process will wait for the server for as long as it takes.